Set lidar data to 16bits from 8

+++ sw/peripherals-board/src/drivers/hwrev2/hwrev2_rf24_shared.hpp
@@ -8,9 +8,9 @@ struct hwrev2_rf24_telem_block1{
   float oriY;
   float oriZ;
 
-  uint8_t lidarLeft;
-  uint8_t lidarFront;
-  uint8_t lidarRight;
+  uint16_t lidarLeft;
+  uint16_t lidarFront;
+  uint16_t lidarRight;
 
 };
