+++ sw/peripherals-board/src/managers/vehicledata.hpp
@@ -13,7 +13,7 @@ struct VehicleData{
   Vec3f orientation;
   Vec3f acceleration;
   Vec3f angularVelocity;
-  uint8_t lidar[360];
+  uint16_t lidar[360];
   int16_t speed;
   long encoderPosition;
